backend/requirements.txt
# FastAPI & server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Settings & validation
pydantic[email]==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-dotenv==1.0.1

# Market data & numerics
numpy==2.2.1
pandas==2.2.3
yfinance==0.2.51

# Background jobs
redis==5.2.1
rq==2.1.0

# Testing
pytest==8.3.4
httpx==0.28.1
